Update write.hpp by using the correct return type

+++ syscall/write.hpp
@@ -1,8 +1,9 @@
 #pragma once
 #include "../fd.hpp"
 #include <stddef.h>
+#include <stdint.h>
 
-[[gnu::naked]] static int write([[maybe_unused]] fd file_des,
+[[gnu::naked]] static int64_t write([[maybe_unused]] fd file_des,
 								[[maybe_unused]] const char* buffer,
 								[[maybe_unused]] size_t count) noexcept
 {
@@ -16,13 +17,13 @@
 }
 
 template<size_t COUNT>
-int write(const char (&buffer)[COUNT])
+int64_t write(const char (&buffer)[COUNT])
 {
 	return write(STDOUT, buffer, COUNT);
 }
 
 template<size_t COUNT>
-int write(fd file_des, const char (&buffer)[COUNT])
+int64_t write(fd file_des, const char (&buffer)[COUNT])
 {
 	return write(file_des, buffer, COUNT);
 }
